+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 ### Changed
+* Changed `Element.set_attribute` args to use `user_attribute_id` as int and `user_attribute` as string.
 
 ### Removed